refactor: standardize panel imports across components
- Updated import paths for the `Panel` component to use the new centralized location in `@lib/ui/panel/Panel`.
- Cleaned up import statements in various files to enhance code organization and maintainability.
- Introduced new `FullPageFlowErrorState` and `MpcPendingMessage` components for improved error handling and user feedback in the MPC flow.
