+++ clients/desktop/src/coin/ui/inputs/CoinOption.tsx
@@ -14,7 +14,7 @@ import { ChainCoinIcon } from '../../../chain/ui/ChainCoinIcon'
 import { useFormatFiatAmount } from '../../../chain/ui/hooks/useFormatFiatAmount'
 import { getChainEntityIconSrc } from '../../../chain/utils/getChainEntityIconSrc'
 import { Skeleton } from '../../../components/skeleton'
-import { panel } from '../../../lib/ui/panel/Panel'
+import { panel } from '@lib/ui/panel/Panel'
 import { shouldDisplayChainLogo } from '../../../vault/chain/utils'
 import { useCurrentVaultCoin } from '../../../vault/state/currentVaultCoins'
 import { getCoinLogoSrc } from '../../logo/getCoinLogoSrc'
