+++ clients/desktop/src/pages/vaultSettings/vaultAddressBook/components/modifyAddressForm/ModifyAddressForm.styles.tsx
@@ -5,7 +5,7 @@ import { MenuProps, StylesConfig } from 'react-select'
 import styled from 'styled-components'
 
 import { Input } from '../../../../../lib/ui/inputs/text-input/Input'
-import { Panel } from '../../../../../lib/ui/panel/Panel'
+import { Panel } from '@lib/ui/panel/Panel'
 
 export const Container = styled.div`
   height: 100%;
