+++ clients/desktop/src/vault/chain/coin/VaultChainCoinPage.tsx
@@ -14,7 +14,7 @@ import { useBalanceQuery } from '../../../coin/query/useBalanceQuery'
 import { getBalanceQueryKey } from '../../../coin/query/useBalancesQuery'
 import { useCoinPriceQuery } from '../../../coin/query/useCoinPriceQuery'
 import { Center } from '../../../lib/ui/layout/Center'
-import { Panel } from '../../../lib/ui/panel/Panel'
+import { Panel } from '@lib/ui/panel/Panel'
 import { PageHeaderIconButtons } from '../../../ui/page/PageHeaderIconButtons'
 import { VaultPrimaryActions } from '../../components/VaultPrimaryActions'
 import { useCurrentVaultCoin } from '../../state/currentVaultCoins'
