+++ clients/desktop/src/vault/keysign/shared/KeysignSigningStep.tsx
@@ -19,7 +19,7 @@ import { CurrentTxHashProvider } from '../../../chain/state/currentTxHash'
 import { TxOverviewPanel } from '../../../chain/tx/components/TxOverviewPanel'
 import { TxOverviewChainDataRow } from '../../../chain/tx/components/TxOverviewRow'
 import useVersionCheck from '../../../lib/hooks/useVersionCheck'
-import { FullPageFlowErrorState } from '../../../ui/flow/FullPageFlowErrorState'
+import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { KeysignCustomMessageInfo } from '../join/verify/KeysignCustomMessageInfo'
 import { KeysignSigningState } from './KeysignSigningState'
 import { KeysignTxOverview } from './KeysignTxOverview'
