+++ clients/desktop/src/vault/reshare/fast/FastReshareServerStep.tsx
@@ -14,7 +14,7 @@ import { useMutation } from '@tanstack/react-query'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { FullPageFlowErrorState } from '../../../ui/flow/FullPageFlowErrorState'
+import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { WaitForServerLoader } from '../../server/components/WaitForServerLoader'
 
 export const FastReshareServerStep: React.FC<OnFinishProp> = ({ onFinish }) => {
